refactoring: refatorando código da classe tarifaService para melhor separacao de responsabilidade das classes e subpacotes no pacote dto
