+++ src/main/java/dev/felipequeiroz/simplepms/controller/CategoriaDeUhController.java
@@ -1,10 +1,9 @@
 package dev.felipequeiroz.simplepms.controller;
 
 import dev.felipequeiroz.simplepms.domain.CategoriaDeUh;
-import dev.felipequeiroz.simplepms.dto.AtualizacaoCategoriaDeUhDTO;
-import dev.felipequeiroz.simplepms.dto.CadastroCategoriaDeUhDTO;
-import dev.felipequeiroz.simplepms.dto.DetalhamentoCategoriaDeUhDTO;
-import dev.felipequeiroz.simplepms.dto.DetalhamentoClienteDTO;
+import dev.felipequeiroz.simplepms.dto.categoriaDeUh.AtualizacaoCategoriaDeUhDTO;
+import dev.felipequeiroz.simplepms.dto.categoriaDeUh.CadastroCategoriaDeUhDTO;
+import dev.felipequeiroz.simplepms.dto.categoriaDeUh.DetalhamentoCategoriaDeUhDTO;
 import dev.felipequeiroz.simplepms.repository.CategoriaDeUhRepository;
 import dev.felipequeiroz.simplepms.service.CategoriaDeUhService;
 import io.swagger.v3.oas.annotations.Operation;
