+++ src/main/java/dev/felipequeiroz/simplepms/domain/UnidadeHabitacional.java
@@ -1,6 +1,5 @@
 package dev.felipequeiroz.simplepms.domain;
 
-import dev.felipequeiroz.simplepms.dto.CadastroUnidadeHabitacionalDTO;
 import jakarta.persistence.*;
 import lombok.*;
