+++ src/main/java/dev/felipequeiroz/simplepms/dto/categoriaDeUh/AtualizacaoCategoriaDeUhDTO.java
@@ -1,13 +1,12 @@
-package dev.felipequeiroz.simplepms.dto;
+package dev.felipequeiroz.simplepms.dto.categoriaDeUh;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 
-public record CadastroCategoriaDeUhDTO(
-        @NotBlank
+public record AtualizacaoCategoriaDeUhDTO(
+        @NotNull Long id,
         String nomeCategoria,
-
         @Min(value = 1, message = "Quantidade mínima é 1.")
         @Max(value = 100, message = "Quantidade máxima é 100.")
         Integer quantidadeDeLeitos) {
