+++ src/main/java/dev/felipequeiroz/simplepms/dto/cliente/AtualizacaoClienteDTO.java
@@ -1,4 +1,4 @@
-package dev.felipequeiroz.simplepms.dto;
+package dev.felipequeiroz.simplepms.dto.cliente;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -6,14 +6,11 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDate;
 
-public record CadastroClienteDTO(
-        @NotBlank
-        String nomeCompleto,
-        @NotBlank
-        String documento,
+public record AtualizacaoClienteDTO(
         @NotNull
+        Long id,
+        String nomeCompleto,
         LocalDate dataDeNascimento,
-        @Email
         String email,
         String telefone) {
 }
