+++ src/main/java/dev/felipequeiroz/simplepms/dto/tarifa/DetalhamentoTarifaDetalhamentoDTO.java
@@ -1,4 +1,4 @@
-package dev.felipequeiroz.simplepms.dto;
+package dev.felipequeiroz.simplepms.dto.tarifa;
 
 import dev.felipequeiroz.simplepms.domain.ClassificacaoHospede;
 import dev.felipequeiroz.simplepms.domain.TarifaDetalhamento;
