+++ src/main/java/dev/felipequeiroz/simplepms/dto/unidadeHabitacional/AtualizacaoUnidadeHabitacionalDTO.java
@@ -1,14 +1,15 @@
-package dev.felipequeiroz.simplepms.dto;
+package dev.felipequeiroz.simplepms.dto.unidadeHabitacional;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-public record CadastroUnidadeHabitacionalDTO(
+public record AtualizacaoUnidadeHabitacionalDTO(
 
+        @NotNull
+        Long id,
         @Size(min = 2, message = "O nome da UH deve ter pelo menos 2 caracteres.")
-        @NotBlank
         String nomeUh,
         @NotNull
         Long idCategoriaDeUh) {
+
 }
