+++ src/main/java/dev/felipequeiroz/simplepms/service/TarifaService.java
@@ -1,18 +1,15 @@
 package dev.felipequeiroz.simplepms.service;
 
-import dev.felipequeiroz.simplepms.domain.ClassificacaoHospede;
 import dev.felipequeiroz.simplepms.domain.Tarifa;
 import dev.felipequeiroz.simplepms.domain.TarifaDetalhamento;
-import dev.felipequeiroz.simplepms.domain.UnidadeHabitacional;
-import dev.felipequeiroz.simplepms.dto.CadastroTarifaDTO;
-import dev.felipequeiroz.simplepms.dto.CadastroTarifaDetalhamentoDTO;
+import dev.felipequeiroz.simplepms.dto.tarifa.CadastroTarifaDTO;
 import dev.felipequeiroz.simplepms.repository.TarifaRepository;
+import dev.felipequeiroz.simplepms.validations.tarifa.CadastrarTarifaValidations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import java.net.URI;
-import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -22,43 +19,25 @@ public class TarifaService {
     @Autowired
     private TarifaRepository tarifaRepository;
 
+    @Autowired
+    private List<CadastrarTarifaValidations> validationsList;
+
     public Tarifa cadastrar(CadastroTarifaDTO dto) {
 
-        // Obter todos os valores possíveis do enum ClassificacaoHospede
-        List<ClassificacaoHospede> tiposEnum = Arrays.asList(ClassificacaoHospede.values());
+        validationsList.forEach(v -> v.validate(dto));
 
-        // Verificar se o número de detalhes fornecidos corresponde ao número de tipos do enum
-        if (dto.tarifaDetalhamentos().size() != tiposEnum.size()) {
-            throw new IllegalArgumentException("É necessário fornecer um detalhamento para cada classificação de hóspede.");
-        }
+        Tarifa tarifa = new Tarifa(dto);
 
-        // Verificar se todos os tipos do enum estão presentes e se não há duplicatas
-        List<ClassificacaoHospede> classificacoesFornecidas = dto.tarifaDetalhamentos().stream()
-                .map(CadastroTarifaDetalhamentoDTO::classificacaoHospede)
-                .distinct()
-                .collect(Collectors.toList());
-
-        if (!classificacoesFornecidas.containsAll(tiposEnum)) {
-            throw new IllegalArgumentException("Cada tipo de classificação de hóspede (CRIANCA, ADOLESCENTE, ADULTO) deve ter um detalhamento.");
-        }
-
-        // Criar a entidade Tarifa
-        Tarifa tarifa = new Tarifa();
-        tarifa.setNomeTarifa(dto.nomeTarifa());
-        tarifa.setValorBase(dto.valorBase());
-
-        // Adicionar os detalhes à tarifa
-        List<TarifaDetalhamento> detalhes = dto.tarifaDetalhamentos().stream().map(detalheDTO -> {
-            TarifaDetalhamento detalhe = new TarifaDetalhamento();
-            detalhe.setClassificacaoHospede(detalheDTO.classificacaoHospede());
-            detalhe.setValorHospedeAdicional(detalheDTO.valorHospedeAdicional());
-            detalhe.setTarifa(tarifa);
-            return detalhe;
+        List<TarifaDetalhamento> tarifaDetalhamentos = dto.tarifaDetalhamentos().stream().map(detalhamentoDTO -> {
+            TarifaDetalhamento detalhamento = new TarifaDetalhamento();
+            detalhamento.setClassificacaoHospede(detalhamentoDTO.classificacaoHospede());
+            detalhamento.setValorHospedeAdicional(detalhamentoDTO.valorHospedeAdicional());
+            detalhamento.setTarifa(tarifa);
+            return detalhamento;
         }).collect(Collectors.toList());
 
-        tarifa.setTarifaDetalhamentos(detalhes);
+        tarifa.setTarifaDetalhamentos(tarifaDetalhamentos);
 
-        // Salvar a tarifa com seus detalhes
         return tarifaRepository.save(tarifa);
 
     }
