+++ src/main/java/dev/felipequeiroz/simplepms/validations/cliente/CadastrarClienteValidations.java
@@ -1,6 +1,6 @@
 package dev.felipequeiroz.simplepms.validations.cliente;
 
-import dev.felipequeiroz.simplepms.dto.CadastroClienteDTO;
+import dev.felipequeiroz.simplepms.dto.cliente.CadastroClienteDTO;
 
 public interface CadastrarClienteValidations {
