+++ src/main/java/dev/felipequeiroz/simplepms/validations/tarifa/CadastrarTarifaValidations.java
@@ -1,0 +1,9 @@
+package dev.felipequeiroz.simplepms.validations.tarifa;
+
+import dev.felipequeiroz.simplepms.dto.tarifa.CadastroTarifaDTO;
+
+public interface CadastrarTarifaValidations {
+
+    void validate(CadastroTarifaDTO dto);
+
+}
