+++ src/main/java/dev/felipequeiroz/simplepms/validations/unidadeHabitacional/NomeJaExistenteValidation.java
@@ -1,6 +1,6 @@
 package dev.felipequeiroz.simplepms.validations.unidadeHabitacional;
 
-import dev.felipequeiroz.simplepms.dto.CadastroUnidadeHabitacionalDTO;
+import dev.felipequeiroz.simplepms.dto.unidadeHabitacional.CadastroUnidadeHabitacionalDTO;
 import dev.felipequeiroz.simplepms.exception.BusinessValidationException;
 import dev.felipequeiroz.simplepms.repository.UnidadeHabitacionalRepository;
 import org.springframework.beans.factory.annotation.Autowired;
