+++ src/test/java/dev/felipequeiroz/simplepms/controller/ClienteControllerTest.java
@@ -3,9 +3,9 @@ package dev.felipequeiroz.simplepms.controller;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dev.felipequeiroz.simplepms.domain.Cliente;
-import dev.felipequeiroz.simplepms.dto.AtualizacaoClienteDTO;
-import dev.felipequeiroz.simplepms.dto.CadastroClienteDTO;
-import dev.felipequeiroz.simplepms.dto.DetalhamentoClienteDTO;
+import dev.felipequeiroz.simplepms.dto.cliente.AtualizacaoClienteDTO;
+import dev.felipequeiroz.simplepms.dto.cliente.CadastroClienteDTO;
+import dev.felipequeiroz.simplepms.dto.cliente.DetalhamentoClienteDTO;
 import dev.felipequeiroz.simplepms.repository.ClienteRepository;
 import dev.felipequeiroz.simplepms.service.ClienteService;
 import jakarta.persistence.EntityNotFoundException;
