+++ src/test/java/dev/felipequeiroz/simplepms/service/ClienteServiceTest.java
@@ -1,8 +1,8 @@
 package dev.felipequeiroz.simplepms.service;
 
 import dev.felipequeiroz.simplepms.domain.Cliente;
-import dev.felipequeiroz.simplepms.dto.AtualizacaoClienteDTO;
-import dev.felipequeiroz.simplepms.dto.CadastroClienteDTO;
+import dev.felipequeiroz.simplepms.dto.cliente.AtualizacaoClienteDTO;
+import dev.felipequeiroz.simplepms.dto.cliente.CadastroClienteDTO;
 import dev.felipequeiroz.simplepms.repository.ClienteRepository;
 import dev.felipequeiroz.simplepms.validations.cliente.CadastrarClienteValidations;
 import org.junit.jupiter.api.DisplayName;
